refactor: update ID types to Long in Compra, Produto, and related controllers and services; enhance CompraProdutoService with new methods for CRUD operations

+++ src/main/java/com/unifunec/vendas/repository/CompraRepository.java
@@ -3,6 +3,6 @@ package com.unifunec.vendas.repository;
 import com.unifunec.vendas.models.Compra;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CompraRepository extends JpaRepository<Compra, Integer> {
+public interface CompraRepository extends JpaRepository<Compra, Long> {
 
 }
